+++ platform/jvm/capture/src/test/kotlin/io/bitdrift/capture/reports/processor/FatalIssueReporterProcessorTest.kt
@@ -18,13 +18,14 @@ import com.nhaarman.mockitokotlin2.verify
 import io.bitdrift.capture.ContextHolder
 import io.bitdrift.capture.ContextHolder.Companion.APP_CONTEXT
 import io.bitdrift.capture.attributes.ClientAttributes
+import io.bitdrift.capture.fakes.FakeDateProvider
 import io.bitdrift.capture.fakes.FakeJvmException
 import io.bitdrift.capture.fakes.FakeLatestAppExitInfoProvider.Companion.createTraceInputStream
 import io.bitdrift.capture.reports.binformat.v1.issue_reporting.Architecture
 import io.bitdrift.capture.reports.binformat.v1.issue_reporting.Platform
 import io.bitdrift.capture.reports.binformat.v1.issue_reporting.Report
 import io.bitdrift.capture.reports.binformat.v1.issue_reporting.ReportType
-import io.bitdrift.capture.reports.persistence.IFatalIssueReporterStorage
+import io.bitdrift.capture.reports.persistence.IIssueReporterStorage
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Before
 import org.junit.Test
@@ -43,24 +44,26 @@ import java.nio.file.Paths
 @RunWith(RobolectricTestRunner::class)
 @Config(sdk = [31])
 class FatalIssueReporterProcessorTest {
-    private lateinit var fatalIssueReporterProcessor: FatalIssueReporterProcessor
     private lateinit var attributes: ClientAttributes
-    private val fatalIssueReporterStorage: IFatalIssueReporterStorage = mock()
+    private val fatalIssueReporterStorage: IIssueReporterStorage = mock()
     private val streamingReportProcessor: IStreamingReportProcessor = mock()
     private val lifecycleOwner: LifecycleOwner = mock()
     private val fatalIssueReportCaptor = argumentCaptor<ByteArray>()
     private val reportTypeCaptor = argumentCaptor<Byte>()
+
+    private lateinit var processor: IssueReporterProcessor
 
     @Before
     fun setUp() {
         val initializer = ContextHolder()
         initializer.create(ApplicationProvider.getApplicationContext())
         attributes = ClientAttributes(APP_CONTEXT, lifecycleOwner)
-        fatalIssueReporterProcessor =
-            FatalIssueReporterProcessor(
+        processor =
+            IssueReporterProcessor(
                 fatalIssueReporterStorage,
                 attributes,
                 streamingReportProcessor,
+                FakeDateProvider,
             )
     }
 
@@ -69,7 +72,7 @@ class FatalIssueReporterProcessorTest {
         val callerThread = Thread("crashing_thread")
         val fakeException = FakeJvmException()
 
-        fatalIssueReporterProcessor.persistJvmCrash(
+        processor.persistJvmCrash(
             FAKE_TIME_STAMP,
             callerThread,
             fakeException,
@@ -108,7 +111,7 @@ class FatalIssueReporterProcessorTest {
                 "OnErrorNotImplementedException",
                 IllegalArgumentException("Artificial exception"),
             )
-        fatalIssueReporterProcessor.persistJvmCrash(
+        processor.persistJvmCrash(
             FAKE_TIME_STAMP,
             Thread("crashing-thread"),
             exception,
@@ -133,9 +136,9 @@ class FatalIssueReporterProcessorTest {
 
     @Test
     fun persistAppExitReport_whenAnr() {
-        doReturn("/some/path/foo.cap").`when`(fatalIssueReporterStorage).generateFilePath()
+        doReturn("/some/path/foo.cap").`when`(fatalIssueReporterStorage).generateFatalIssueFilePath()
         val trace = buildTraceInputStringFromFile("app_exit_anr_deadlock_anr.txt")
-        fatalIssueReporterProcessor.persistAppExitReport(
+        processor.persistAppExitReport(
             fatalIssueType = ReportType.AppNotResponding,
             enableNativeCrashReporting = true,
             FAKE_TIME_STAMP,
@@ -156,7 +159,7 @@ class FatalIssueReporterProcessorTest {
         val description = "Native crash"
         val traceInputStream = buildTraceInputStringFromFile("app_exit_native_crash.bin")
 
-        fatalIssueReporterProcessor.persistAppExitReport(
+        processor.persistAppExitReport(
             ReportType.NativeCrash,
             enableNativeCrashReporting = false,
             FAKE_TIME_STAMP,
@@ -172,7 +175,7 @@ class FatalIssueReporterProcessorTest {
         val description = "Native crash"
         val traceInputStream = buildTraceInputStringFromFile("app_exit_native_crash.bin")
 
-        fatalIssueReporterProcessor.persistAppExitReport(
+        processor.persistAppExitReport(
             ReportType.NativeCrash,
             enableNativeCrashReporting = true,
             FAKE_TIME_STAMP,
@@ -221,7 +224,7 @@ class FatalIssueReporterProcessorTest {
         val description = null
         val traceInputStream = createTraceInputStream("sample native crash trace")
 
-        fatalIssueReporterProcessor.persistAppExitReport(
+        processor.persistAppExitReport(
             ReportType.JVMCrash,
             enableNativeCrashReporting = true,
             FAKE_TIME_STAMP,
@@ -247,9 +250,5 @@ class FatalIssueReporterProcessorTest {
 
     private companion object {
         const val FAKE_TIME_STAMP = 1241515210914L
-        const val APP_EXIT_DESCRIPTION_ANR =
-            "Input dispatching timed out (219180 " +
-                "io.bitdrift.capture/io.bitdrift.capture.MainActivity (server) " +
-                "is not responding. Waited 5004ms for MotionEvent"
     }
 }